Services don't use Spring classes
